requirements.txt
pandas
mlflow
notebook
scikit-learn
numpy
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
Flask
Flask-Cors
types-PyYAML
-e .